fix(ESSNTL-5110): Improve RBAC checks for delete and edit button on the System details view (#1948)
* fix: Disable buttons when not enough permissions

This sets the minimal permissions required in order for users to be able
to press Delete and edit buttons on System details page. If a host is in
a group, then it also checks if user has group level access (which is a
softer reqirement compared to general inventory:hosts:write).

* fix: Edit modal, change buttons order and type

* fix: Remove permissions cache for edit buttons

* test: Test the delete and edit buttons behavior

* test: Update tests for EditButton